Block Selection Form Updates...
Changed BlockStates CheckedListBox to bind to a DataView, to allow filtering its contents

Added a TextBox for use in filtering the BlockState CheckredListBox to items matching the text entered.

Replaced [Select All] and [Unselect All] buttons with a CheckBox that operates on the currently displayed items.

Made the current image update to the first selected variant anytime the BlockVariant CheckedListBox is updated.  So when moving through the BlockState ChecedkListBox, the image updates according to whichever BlockState is selected.

Added split between BlockStateVariant ChecedkListBox and the image, to allow resizing the BlockStateVariant CheckedListBox.

Updated the form's layout and anchor settings, so controls resize appropriately.